Modifications to doc site
